# Docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Next.js API (Feature Requests, Profiles, Personas, Contentstack)",
    "description": "API for authentication (Supabase), feature requests (with comments/votes and file uploads), user profiles & personas, Contentstack publishing helpers, and internal log ingestion.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/feature-requests/{id}/vote": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "description": "Feature request ID"
        }
      ],
      "post": {
        "tags": [
          "Feature Requests"
        ],
        "summary": "Vote on a feature request",
        "description": "Upserts or removes the current user's vote for a feature request and updates the `number_of_votes` aggregate. Requires `COMMENT_VOTE_FEATURE_REQUESTS` permission.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vote recorded; updated vote total returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid vote value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (missing permission).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "operationId": "postApiFeatureRequestsIdVote"
      }
    },
...
    "/api/personas": {
      "post": {
        "tags": [
          "Personas"
        ],
        "summary": "Create a persona and set it active",
        "description": "Creates a persona owned by the authenticated user and sets it as the user's active persona. Requires `USE_PERSONAS` permission.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonaRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Persona created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Persona"
                }
              }
            }
          },
          "400": {
            "description": "Persona name required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (missing permission).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "operationId": "postApiPersonas"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-<project-ref>-auth-token",
        "description": "Supabase SSR session cookie (project-specific name like `sb-<project-ref>-auth-token`)."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer access token (e.g., Supabase access_token)."
      },
      "internalLogSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-log-secret",
        "description": "Shared secret used to authorize internal log ingestion."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "error"
        ]
      },
...
      "VoteRequest": {
        "type": "object",
        "required": [
          "vote"
        ],
        "properties": {
          "vote": {
            "type": "string",
            "enum": [
              "up",
              "down",
              "remove"
            ]
          }
        },
        "additionalProperties": false
      },
      "VoteResponse": {
        "type": "object",
        "properties": {
          "number_of_votes": {
            "type": "integer"
          }
        },
        "required": [
          "number_of_votes"
        ],
        "additionalProperties": false
      },
...
      "Persona": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid"
          },
          "full_name": {
            "type": "string"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "avatar_url": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": true
      },
      "CreatePersonaRequest": {
        "type": "object",
        "required": [
          "displayName"
        ],
        "properties": {
          "displayName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}